feat:added  def test_TC_005_09_01_verify_solar_irradiance_description(driver):
