+++ tests/test_jenya_solar_irradiance/pages/solar_irradiance_page.py
@@ -1,0 +1,21 @@
+from pages.base_page import BasePage
+from locators.locators import SolarIrradianceLocators
+
+
+
+
+class SolarIrradiancePage(BasePage):
+    locators = SolarIrradianceLocators()
+
+
+    def verify_description_visibility(self):
+        locator = self.locators.SOLAR_IRRADIANCE_DESC
+        element = self.driver.find_element(*locator)
+        assert element.is_displayed()
+#check description visibility
+
+# check link hover
+
+# check link clickability
+
+#check table visibility
